+++ source/program.h
@@ -1,0 +1,8 @@
+#ifndef PROGRAM_H
+#define PROGRAM_H
+
+// Should be in socket.h
+int
+inet_pton(int af, const char *src, void *dst);
+
+#endif /* PROGRAM_H */