+++ src/program.h
@@ -1,6 +1,0 @@
-#ifndef PROGRAM_H
-#define PROGRAM_H
-
-int _entryPoint(int argc, char **argv);
-
-#endif /* PROGRAM_H */